Adds gatestringWeights argument to MLGST.
Now `do_mlgst` has a 'gatestringWeights' argument and
`do_iterative_mlgst` has a 'gatestringWeightsDict' argument, similar
to the corresponding MC2GST routines.  Setting these arguments
specifies multplicative prefactors for the logL terms on a by-
gatestring basis.  This functionality was recently useful in a
MC2GST anlaysis and so it seemed worthwhile to add to to MLGST.