Adds files for new report framework.  Not usable yet.
This commit adds several files to the report subpackage and
the new "algorithms/directx.py", as the report code is being
refactored and rewritten for plotly/html instead of Latex/pdf.

The new code centers around the Workspace object which is able to
create plots and tables in an interactive setting while caching
intermediate calculations.  Eventually the workspace*.py functions
will replace plotting.py and generation.py (and maybe more).

Curently the structure of the code is taking form but there's still
much left to do and this new code should be considered NON-FUNCTIONAL
at this point.