Adds core Switchboard logic & initiates widget->HTML transition.
This commit transitions from the "widget" model of workspace
outputs, whereby:

- each output object contains a single "current" state and display()
  shows a widget with this state.
- update() is called to re-render the widget, and it is the
  responsibility of the Selector / Switchboard to call update() on
  all relevant outputs.
- interaction is performed via ipywidgets

to the new (and better, I think) "HTML" model whereby:

- each output object holds an entire range of results as dictated
  by its connection to Switchboards.
- render() creates HTML + javascript necessary for all interaction
  with said Switchboards (which integrates with the HTML+JS produced
  by Switchboard.render()).
- display() just renders and places the result in an HTML widget.
- it is the output object's responsibility to listen for relevant
  switch events.
- interaction is performed via JQueryUI javascipt components.

This significant change has resulted in a new switchedCompute(...)
method which is similar to cachedCompute(...) but returns *all* of
the precomputed results along with a mapping to switch positions.
In the future these two methods should be consolidated.

The motivation behind this shift stems from the following
reasonings about the new model:
- ipywidgets don't offer very much logic that's especially hard to
  do without them, and they're specific to ipython notebooks.
- the new HTML+JS model can be displayed in standalone web pages
  or within a notebook.
- the interaction model is more extensible and cleaner, thanks to
  JQuery's  allowance of multiple event handlers.
- interaction seems nearly as flexible as using widgets, and works
  with tables and plots in the same way -- and maybe is faster?
- probably more I'm forgetting right now.

At this point, WorkspaceValue and derived classe are still present,
but seem perfunctory and will likely be removed soon.  The code is
still in a very rough and fragile state, and many features don't
quite work yet.  Among other things, Switchboards and WorkspacePlots
can render themselves in HTML but tables can't yet; Switchboards
don't start with initial values; there are hardcoded IDs for
debugging; only toggle/radio switch type works.