Adds init_notebook_mode to Workspace & enables HTML saving in Jupyter.
Adds an init_notebook_mode function to Workspace for performing all
the needed initialization in an ipython notebook.  The tricky part
is getting Jupyter to save a working HTML file, which is now
nomially working.  Since Jupyter uses require.js (and includes this
when downloading as HTML) one must set the "global_requirejs" flag
hidden in the depths of Plotly.  This flag isn't exposed in the
public API (but really should be), so this commit adds a the
"plotly_offline_fixed.py" file with duplicates plotly.offline.plot
but with an additional (optional) "global_requirejs" argument.
This may cause a headache as plotly gets updated, but it's a working
stopgap solution.

Also in this commit, the displayed tables, plots, and switchboards
are no longer HTML ipywidgets.  This is because after updating my
Jupyter notebook (to get the HTMLMath widget, which worked) the
HTML widgets stopped allowing execution of javascript.  This is fine,
since we can use display(HTML(x)) which still runs HTML and JS as
expected.