Adds assume_sorted=False to TDDataSet in hopes of fixing TravisCI tests.
TravisCI tests are failing at Scipy's interp1d routine with an error
'ValueError: Expect x to be a 1-D sorted array_like.' but this is not
seen on my local machine.  This could be because the x-data using in
the test is not monotonically increasing (but has some successive
equal values).  Since this logic will be replaced in the near future
anyway it doesn't seem worth it to try to debug what's upsetting
scipy at this point, so this commit just adds "assume_sorted=False"
to the interp1d calls in hopes this will fix the issue.