Updates GatestringStructure usage; adds do_long_sequence_gst_base.
This significant update adds docstrings to gatestringstruct.py and
gives the GatestringStructure and LsGermsStructure classes well-
defined meanings and carries their usage throughout pyGSTi.  Now
`do_long_sequence_gst` no longer has a "lsgstLists" argument;
instead there is a new function, `do_long_sequence_gst_base` which
takes a list of either raw gate string lists or of LsGermsStructure
objects, which conveys all of the gatestring structure information
for running LGST, LSGST, and initializing a Results object.

A "gate string structure" is now understood to be a set of gate
sequences (similar to a list of GateStrings) with additional
structure used for LGST and post-processing/visualization.  Each
structure is composed of "plaquettes" corresponding to the sub-pieces
of a color box plot.  A new GatestringPlaquette class encapsulates
this and does away with the messy "gatestring_map" functions throughout
the plotting code, which is much more readable now.