Fixes some bugs and adds some minor enhancements to HTML reporting.
Fixes bugs in LsGermsStructure.copy() and with updating sliders
when the sliders allowed values are not the same as the corresponding
switch indices.

Colormaps have been updated to have 'hmin' and 'hmax' values which
specify the "heatmap-min" and "heatmap-max" values that should be
given to plotly's Heatmap function.  For linear mappings, these can
(and should) be the min and max values, which plotly then automatically
interpolates (linearly) onto the [0,1] range.  For Linear-Log color
maps, however, normalize() maps the values to the [hmin,hmax] == [0,1]
range first and then plotly trivially maps this to [0,1] for color
determination.  This fixes the "besttxtcolor" function so that box label
colors are now correctly determined.

Also adds a "scale" parameter to ColorBoxPlot and makes the Polar
plots look a little nicer.