Adds "alwaysPerformMLE" option for do_iterative_mlgst and drivers.
When this new argument of do_iterative_mlgst is set to 'True',
do_mlgst is called after (not instead of) the usual call to
do_mc2gst on *every* iteration of the parameter estimation.
Previously (and by default) MLGST is performed only after the
*final* iteration.

The driver routine do_long_sequence_gst also has an added
advanced option, 'alwaysPerformMLE' which pipes directly into the
call to do_iterative_mlgst.