Swaps out faulty numpy linear solver for scipy's.
Erik recently observed that numpy.linalg.solve will sometimes
silently fail (no exceptions raised), which subsequently causes a
failure of pyGSTi's LM algorithm to converge (as it depends on the
linear solver telling it when the system is singular).  Swapping
in scipy.linalg.solve, which uses a more tailored (assumes symmetric
and positive 'A' matrix) and, apparently, robust numerical method,
seems to have fixed the problem.

In the 2Q-GST test case where this was observed, the numpy solver
would continually give larger-norm solutions when A's diagonal
was made increasingly large (this shouldn't happen).  Scipy's
solver correctly raised a LinAlgError (singular system) which is
handled by the LM routine and ultimately leads to a convergence
condition.