Adds initial capability to DataSet for intermediate measurements.
Adds the abilityt to specify a dictionary (as the
'measurementGates' argument of DataSet or load_dataset) of set of
gates which should be treated as the operation performed by a gate
that has classical outputs contingent on one of the outcomes.
Counts from sequences which differ only in Gates which correspond
to different outcomes of the same gate-with-classical-output will
be summed when computing the total number of counts for any of the
sequences.  This summation is done entirely within DataSet, so
outside of this performing GST on gates that have classical outputs
will look like performing normal GST, but with one Gate per
*outcome* of the gate-with-classical-output.