Updates auto-sizing and resize functionality, and adds Katex.
Neatens up the auto-size and resizable-widget functionality to the
point where both can almost be used concurrently if it makes sense
to do so.  Work still needs to be done on fluid layout to get things
to look correct, so current default is to just make plots resizable
but *not* auto-sizing.

This commit applies a more major fix which essentially makes polar
plots resizable (plotly doesn't support this for some reason).

Lastly, Katex is now used instead of mathjax to perform the math
rendering.  It's much faster.  Mathjax must still be used in plotly
if we want math expressions in plots, but for now they've been
removed so we don't need mathjax at all.