Two significant updates to dashboards.
First, this commit "upgrades" the way fluid plots and tables
are done, making entires tables resizable (instead of the
individual cells) and moving to a more common treatment of
plotly plots (as a *floating* absolute-positioned div inside
an "relwrap").  The allowed and automatic sizing is not limited
to the "natural" (initial) size of the plots, so that plots don't
get too large when expanding the screen.  We may want to allow
manual resizing to remove this constraint, but this is left as a
future enhancement.  Currently, each plot has a parent "container"
(not its immediate parent: for a table cell plot, the TD; otherwise
 the "group" div of all the plotly plots related to a single logical
 plot), and upon resizing the plotly plot changes its size based on
its container.  The element that is made "resizable" is not necessarily
the container (it's a DIV containing the logical table in the table
 case).  Due to the many attempts to get the fluid layout working
well, there's lots of commented out code related to resizing that
need to be removed.  That said, the basic fluid-layout functionality
seems to be working decently well.  The reports now use make figures
auto-resizing and resizable widgets, as these play nicely together
now (elments will auto-size with viewport changes until they are
manually resized, then they no longer auto-size).

Secondly, this commit adds better "connected=False" functionality,
and initiates the presence of an "offline" folder beneath
reports/templates which holds all the resources needed to
display reports and notebooks offline.  The idea is that a copy of
the offline folder will need to be in the same directory as an
"offline-mode" HTML file for things to work.  I looked into options
for embedding all the offline resource content into the HTML file
itself, but this is difficult for images and fonts, and though there
could be workarounds the current approach seems much cleaner.
There's still much to do, but the offline folder should be complete
(unles we decide to add in mathjax) and offline viewing of reports
seems to work after you manually copy the offline folder.