Further improves plotly plot loading so no initial resize is needed.
Previously, plotly plots were created using standard plotly JS and
then resized later.  This was slower than it needed to be, since
resizing takes time comparable to plot creation.  Now, but extracting
the JS from the standard plotly HTML+JS block, we can defer plot
creation until the document is ready and the appropriate initial
size is known -- so now plots are just created once with the correct
initial size, resulting in faster page loading.

A side effect of this is that the JS associated with a single
plotly plot is just one <script> ... </script> block, making the
HTML/JS code a bit cleaner.