Adds several enhancements for placing Switchboards to reports.
1) Use of a flagging CSS class ("initializedSwitch"), Switchboards be
created before or after their dependent output objects, as output
objects will now wait (via polling) for all of their switches to
be initialized before connecting to them.

2) An initial version of master/slave "scale locking" has been added,
allowing the different plotly plots in a single logical plot (i.e.
WorkspacePlot) to be scaled based on the scaline of one designated-
master plot.  I was thinking this would be useful for the color box
plots of different sizes, but it turned out not to because of the
x/y-axis space that is needed for all box plot sizes.  Maybe this
functionality can be augmented later to deal well with this case
(maybe scale based on the actual plot-region sizes by querying the
SVG objects?) but it's not a critical feature so it's being left
as-is for now.

3) A new SwitchboardView class allows the creation of multiple "views"
of the same Switchboard, which will likely be useful in reports and
perhaps notebooks.  Views will wait for their "base" switchboard
to initialize in a page, and so can be displayed before or after
their base.  Views can be restricted to only a subset of the base
Switchboard's switches, so that different controls of the same
board can be displayed at different (& multiple) places in a
page.

4) "numslider" has been added as a new and distinct switch-type.
Previously, "slider" decided whether to use numerically-spaced
positions based on whether it's allowed values were numeric. As
this sometimes lead to unwanted behavior, now the user has
complete control over whether the slider points are positioned
according to their numeric values (a "numslider") or just
equally spaced (a "slider").

5) In the course of testing, Switchboards have been added to the
single-qubit full report.  This has added naturally the ability
to run a single-qubit report on multiple Results objects (which
can be selected via a dropdown switch).  Finally, minor CSS updates to the JqueryUI
controls were made to improve the appearance of the controls.