Improves offline-mode capabilities, especially for notebooks.
The "offline" folder is now automatically updated with newer files
 (when necessary), instead of only being copied when there is no
 offline folder at all.

Some intricacies of requireJS have been figured out - namely that
the base URL is different when running in a notebook vs. in a
saved-as-HTML notebook, and `Workspace.init_notebook_mode` has been
substantially updated to account for this and ensure that in both
cases requisite offline content is located properly.  Additionally,
running init_notebook_mode with connected=False now injects the new
'jupyterlibload.js' file into the notebook, which will pull in offline
resources that Jupyter's HTML assumes will be available via an online
connection.

All together, these updates should result in the files created with
'connected=False' being unreliant on any external resources (though
they will need a local offline/ folder).  The exception is MathJax,
which is assumed to be available by Jupyter but not kept as a local
pyGSTi resource (and so cannot be loaded by jupyterlibload.js).

Finally, when 'connected=True' is used, pyGSTi-specific content is
now inserted *inline* so that an offline/ folder is not needed, as
is desirable (since we don't stage offline files when connected=True).