Cleans up workspace plot & table JS, making page loading more robust.
The HTML and JS for plotly plots are now kept separate, and all of
the JS for a workspace plot or table is aggregated into a *single*
<script> block and placed within a requireJS (when desired) and
ready handler.  This cleans up the HTML and avoids race conditions
which had been causing problems with plots not loading in the past.
Also, much of the JS logic has been pushed to functions within
pygsti_plotly_ex.js, cleaning up the Python code and making it
easier to edit the JS.  Comments have been added indicating the
method by which plots are initialized (sizes set) and created (actual
plotly JS calls) to support a resizable layout.

In addition to cleaning up the plot creation code and making things
generally more robust, this commit adds a check for MathJax in the
chain of table-loading logic that typesets the table before locking
down the widths of non-plot columns (since a rendered math formula
can have a much different shape than its source Latex code), as this
was causing a weird wrapping of formulas in certain cases.

Note: the case with non-resizable plots hasn't been tested much, and
should be in the future.

Finally, when loading offline jupyter-saved-HTML files and MathJax
is unavailable, logic from init_notebook_mode will render the in-page
math using Katex (since pyGSTi has an offline copy of Katex).