Adds usage of custom plotly in notebooks & adds status indicators.
A variant of plotly's get_plotlyjs function is now used to perform
Plotly's init_notebook_mode initialization with our fixed and
improved version of plotly.  Interestingly, only minified versions
work for inserting into notebooks (when connected=False), so this
commit adds a minified (via jscompressor.com) version of our
plotly-polarfixed.js.

Secondly, pyGSTi's init_notebook_mode now displays status information
to confirm which JS libraries have been successfully loaded.

Finally, a stray backslash in a "<\/script>" tag was removed.  In a
previous commit, this was needed as it created a JS error when is
was *not* included - but maybe that was when it was in a .js file(?).
In this case, the backslash was removed from a python string, and
seems to have fixed an issue with pygsti_plotly_ex.js loading in
saved notebook HTML files.