Updates "data comparator": slight refactor and now uses Plotly
The "data_comparator" class which was in reports/datasetcomparison.py
has been renamed to "DataComparator" and relocated to
objects/datacomparator.py for better organization & convention
conformance.  The matplotlib data-comparison plotting routines have
been removed from DataComparator and recreated (to the best of my
ability) using Plotly in a new DatasetComparisionPlot class and as
a new "dscmp" display option of ColorBoxPlot.

The demo notebook has been updated and renamed to Tutorial #23
since the dashboards merge added Tutorials 20-22.  Docstrings still
need to be added throughout datacomparator.py and for the new
DatasetComparisonPlot.