Updates report-containing tutorials to use new reports.  Fixes bugs.
Now all the tutorial notebooks should be using the new reports.  In
converting the gate string / fiducial reduction tutorials a bug was
discovered in which cause color box plots to show up blank when using
reduced gate sequences (due to using numpy.max instead of numpy.nanmax,
which caused the vmax given to the linlog colormap to be nan).

This has been fixed now, along with a bug in 'make_lsgst_structs' which
caused the specification of fiducial pairs to be ignored when keepFraction
was left to 1 (the default).

An issue to note is that when loading gatesets from text files the
default_gauge_group isn't set.  It probably should be (in the future).