Updates reports to work with restructured Results objects.
Report functions now handle multi-estimate Results objects, and
automatically create a switch (a dropdown) for different estimates.

Previously, one could pass a dict of Results objects with the same
gate sequences to the report functions -- this functionality has
been removed since it was just a temporary measure waiting for the
multi-estimate Results now implemented.  That said, it would be
useful to allow a dict of Results objects to create *another*
switch and not require the same gate sequences, dataset, etc, for
these.  To see how we dealt with dicts of Results pull the commit
before this one.