Improves gatematrix box plot sizing and handling in gate mx table.
In particular, the CI column of the gate matrix table, when the
matrix type is *not* a Full or TP gate, now creates a long column
of boxes that shows up instead of being too skinny to see.  The
sizing of the extreme aspect ratio plots still could use more
improvement, but this is better.