Fixes issue with max-height CSS for workspace tables.
Previously, the max-height set for divs underneath figures could
cause issues when a table's TD heights make it taller than this
max height.  Since HTML tables are not block objects, the table
just overflows the text beneath the containing div, and looks bad.
Now when a table's plots are initialized the div's max-height is
removes (set to "none") so that the div expands to fit its contents.
This means that tables can be larger than this max-height, but
this seem to be the least of the evils.