Fixes a bug related to confidence regions in reports.
The "cri" switchboard value was improperly initialized to the
"estimate" and "L" switches when it should have been associated
with the "estimate" and "gauge opt" switches.  This bug could
result in various assertion failures during report creation with
CIs.  Fixed now.

Also, this commit begins to implement in-report switching based
on goodness of fit, e.g. the foundation for switching to non-Mark.
CIs when the fit is bad.