Adds new "data-downscaling" implementation of non-Markovian error bars.
The do_long_sequence_gst(...) driver will now create an additional
Estimate by down-scaling the counts of the DataSet when a bad fit
is obtained.  This is turned on by default, and is controlled by
the advanced options "badFitThreshold" (# of sigma to consider a
"bad fit") and "onBadFit", which defaults to "scale data" but can
also be set to "do nothing" or "scale data and reopt" (which just
raises a NotImplementedError for now).

Chi2 and likelihood functions have been updated to accomodate this;
namely *_terms functions have been added to allow computation of the
log-liklehood or chi2 on a per-gate-string basis as is needed to
assess which gate strings' data should be downscaled.

Report functions have an added "nmthreshold" argument to set the
threshold at which (old-style) non-Markovian error bars are computed,
and the reports will automatically generate plots for all of the
Results' Estimates, so the data-downscaled case is handled nicely.
A new scaling-factor box plot has been added to indicate how
much each gate sequence was scaled.