Adds plot creation and math typeset throttling to dashboard report.
Plot creation and Katex rendering is now queued and executed with
timer-based throttling.  A new Javascript object, PlotManager
still needs to be updated to process its queue in the correct order,
but other than this it seems to work.

Table and figure captions have also been broken into a "title" and
"details" portions, with only the details part being toggled when
clicking the plot (only applied to dashboard report template).