Fixes minor formatter bugs.
convert(...) function squeezed arrays in item_type() call but
not for object being converted - so sometimes item_type would say
a numpy array (e.g. with shape (N,1)) was a "vector" but rendering
wouldn't work.  Now squeezing is done in convert() and not in
item_type().

Also adds 2nd parameter to 'GateString' formatter lambda calls, as
I *think* this is needed -- please check this.