Fixes a bug in hide/show javascript logic.
Previously, ".is('visible')" was used to check if an element had
"display: none" set on it, but this is incorrect, as an element
may be invisible just because a parent has display: none set.  Now
we correctly check for "display: none" using ".css('display')".  I
think this bug may be responsible for the seemingly random
vanishing of plots in the reports.