Improves gate decomposition table, though not fully correct yet.
Gate decomp table which lists gate angle and axis has been broken
for > 1Q gates.  This turned out to be, at least in part, due to
ambiguity in the matrix log of 2Q gates which have multiple -1
eigenvalues.  The real_matrix_log function successfully computes
real logarithms for such gates, but this may not be exactly what
we want, since these real logs do *not* project onto the standard
Hamiltonian errors as expected for standard 2Q target gates.  It
seems the ambiguity in the log can cloud the interpetation of the
error generator projections.  By taking a standard (consistent
branch cut) log the projections of the standard gates again make
sense, so this commit switches to this method.  However, there
remains issues with 1) the normalization used when computing the
"angle" of a gate, and 2) if & how we display *complex* logs of
gates (we could just use the magnitude, but then we lose sign info),
and 3) using the standard log seems bad for CNOT&CPHASE gate
interpretation, as all H projections are 0 in this case, whereas
they're meaningful values when using the real log.