Switches error_generator back to using real_matrix_log(...)
I had switched this to using custom_matrix_log after discovering
that real_matrix_log didn't work correctly in certain cases.  The
thinking is now that real_matrix_log *should* work with appropriate
modifications, and should only (currently) fail for cases when the
matrix being log'd is far from the identity, which isn't the case
in at least the logTiG-type generator.  So I'm reinstating
real_matrix_log within error_generator(...) with plans to debug
and fix it in the future.