Changes fallback eigenvalue clipping from 1e-15 to 1e-12.
When the Cholesky decomposition of a positive matrix is taken
sometimes due to numerical roundoff (I think) issues a numpy
LinAlg error is thrown saying that the matrix is not positive.
To work around this, we fall back to clipping the (all
non-negative) eigenvalues with a LindbladParameterizedGate
construction to some minimum tolerance.  This tolerance used to
be 1e-15 but we are still getting errors in some cases, so I'm
raising it to 1e-12.