Rearranges standard 2Q germs so Gii is first; adds std2Q_XXII.py
A slight re-ordering of the germs in some of the standard 2Q
germ sets so that Gii is the first germ, as this seems like a
sensible convention.  Also adds a new "XXII" gateset which
includes Gii, Gix, Giy, Gxi, Gyi, and Gxx.  Known by some as
a "crosstalk gateset", since it consists of a minimal set of
of gates to do 2Q crosstalk detection within the 2Q GST
context.