Temporarily disables SPAM vecs in Gauge opt Jac.
In the least_squares routine, all of the S matrices
(GaugeGroupElements) have identity matrices for their transforms.

However, since the derivative of a prep vector is - dS_inv/dv (dot) P (?),
The analytic Jacobian holds non zero values.

Its possible this derivative is wrong (As, obviously, it should be zero when
the gauge group elements are identity matrices.)

Another possibility is that the derivative is right but that there is a bug
in the code that calculates it. I'll look into this.

I found no issues with the gateset transform methods,
as the prep elements in both the original, transformed, and target gatesets are
exactly the same throughout the gauge optimization process (because
all of the gauge group elements are identity matrices)