Delays rendering report's logl scatter plot to avoid WebGL issues.
It seems that, at least sometimes, the WebGL-powered logL scatter
plot doesn't render correctly in a browser when it's queued and
created along with the page load.  As a (hopefully temporary)
workaround, this commit marks the plot for delayed creation, so
that a user click is required to trigger plot creation.  The
plot seems to always display properly now.