Updates LindbladParameterizedGate's transform() function to avoid logs.
Now LindbladParameterizedGate.transform directly conjugates it's
unitary prefactor and exponent and so avoids the roundabout way of
re-initializing an entirely new gate (which was especially
problematic given the logarithm ambiguity issues we've discovered
recently).  The new transform() has been "checked" by running
both the old and new implementations and ensuring they give the
same results.  Currently the old implementation is left but
commented out, and should be removed in the future.

At a technical level, this required factoring out a part of
LindbladParameterizedGate's __init__ method into a a new
_set_params_from_errgen(...) function which is called
separately from within transform(...).