Updates & adds missing sequence actions to do_long_sequence_gst.
The make_lsgst_structs function now has a `actionIfMissing` argument
which can be set to "raise" or "drop".  do_long_sequence_gst supplies
"drop" as the default argument, meaning that generated sequences
that are not in the `dscheck` DataSet argument are not included
in the returned structures and therefor not included in the GST
analysis.

Also added to make_lsgst_structs is a `verbosity` argument to
control how much detail regarding the missing sequences is output.