Updates and moves gatestring pre-processor manipulation routine.
The "apply_seq" function that peforms replacements in gate strings
according to rules has been moved to gatestringconstruction.py as
the new `manipulate_gatestring` function.  The logic of the function
has been somewhat reworked to allow for find,replace pairs that are
of unequal lengths.  The gatestringtools.py file has been removed,
as apply_seq was the only function within it.