Finishes initial plumbing of gate string manipulation via rules.
do_long_sequence_gst now accepts a new 'stringManipRules' advanced
option which sets downstream `sequenceRules` arguments.  This
propagates to the creation of a LsGermsGateStringStructure object
which stores the rules, and applies them to *only* it's `allstrs`
member.  I'm not sure if this is really what we want - we'll have
to do some use testing - but in any case it should result in the
`manipulate_gatestring`'d sequences to be sent to the main GST
analysis.