Fixes a bug in color box plot hover labels (measure-fid was flipped).
The logL, etc, color box plots had hover labels which incorrectly
flipped the measurement fiducial that was shown in the on-hover label.
For instance, the *upper* left corner of each plaquette shows the
first (usually trivial) prep & measure fiducial case, and this was
being displayed in the hover info at the *bottom* left corner.