Updates tutorial notebooks; adds Examples and FAQ.
Tutorials have been restructured and updated to contain the most
pedagogical material, and an Examples directory has been started
with notebooks which are meant to illustrate a particular task.
Additionally, a FAQ.ipynb notebook has been started (currently just
a skeleton) to point users to resources by answering frequently asked
questions.

Currently the Tutorials are almost complete (only the Basis tutorial
needs to be written), the Examples will run (except for Model
Selection.ipynb) but more should be added, and the FAQ is
incomplete.