Fixes a caching bug that caused switches to malfunction on some report plots.
Inside the switchedCompute function the hash key for a function and
its arguments is sometimes not computed because this is inefficient.  In
these cases, however, the same "NA" key was used as the "key", and this
caused the logic to re-use the first item created and ("link" the 2nd and
following switch positions to the first).  This commit changes this
key to "INEFFECTIVE" and adds a conditional clause so that an
"INEFFECTIVE" key is alwasy treated as a *distinct* computed item.