Changes minimum k=Ns-Np value to 1 instead of 0 to avoid division by 0.
In several places we compute the goodness of fit related to a
chi^2_k distribution, where k=Ns-Np is the difference in the number
of maximal-model parameters (Ns) and the number of GateSet
parameters (Np).  In the rare cases when Np >= Ns we used to set
k = 0, but this is problematic since we divide by k (or sqrt(k)) to
compute the number of standard deviations of misfit - so now this
minimum is 1, and a warning is issued when Np >= Ns.