Adds auto-uuid generation to old static DataSets.
When a DataSet is static is now needs to have a UUID (for hashing).
Previously this defaulted to None for old datasets that don't have
a 'uuid' member, even if they are static but this is incorrect.  Now
a new UUID is generated for static DataSets when they are unpickled
and don't have a 'uuid' member.