Allows basis_matrices function to take a Basis object as input.
Previously, basis_matrices would only accept a basis name and
dimension.  Given that we'd like to move to utilizing Basis objects
wherever possible, it's helpful to have this routine also accept a
Basis object - in which case "generating" the basis matrices is
trivial since they're contained in the Basis object.