Moves gate decomposition table logic to reportables and adds framework.
In order to get error bars on all of the items in the gate/gateset
decomposition table, the computations within DecompTable have now
been moved to the `general_decomposition` function in reportables.py.

Since the table involves angles between the rotation axes of
different gates, general_decomposition is a function of *gate sets*
and a new gatesets_quantity decorator has been made to generate
error bars in this case.

Also, all of the *_quantity decorators have been updated (along
with their corresponding get_*_fn_confidence_interval functions) so
that the computation functions can now return a *dictionary* of
floats or numpy arrays.  In this case, error bars are computed for
each dictionary value separately, and the *_quantity decorators
will morph the computation function so it returns a dictionary of
ReportableQty objects with the same keys as those of the dictionary
returned by the original computation function.