Adds shortcut to general_decomposition(...) reportable.
Now when general_decomposition computes a gate matrix log, it will
only use the slow approximate_matrix_log function when the target
gate has at least one -1 in its spectrum.  I think this is the only
time when the logarithm is ambiguous in a way where we'd rather
compute an inexact log that is close to the target instead of
the exact log.