Adds create_report_notebook(...) function to make an "advanced report".
This commit finished up the work incorporated by the recent pull
request -- now factory.py contains a create_report_notebook function
which builds an jupyter notebook that, after the user runs it, will
create many of the plots (and text) of the general HTML report.

This "advanced report" functionality is still rough, but it does
run.  A unit test has been added which generates a sample report
notebook.