Cleans up report generation function signatures; adds auto-titling.
The 'datasetLabel' argument has been removed from create_general_report
and create_report_notebook functions, and their 'title' argument has
become more mandatory.  If a title is not supplied (i.e. if the the
default "auto" is left unchanged) then a random report title is
now generated using some mad-libs word lists Erik found on the
internet, and a warning is issued.  Tutorial and example notebooks
have been updated to include 'title=' specification in report
generation calls.

This commit also polishes the notebook report some more, so it
actually uses its 'title' argument now, and the 'connected' argument
gets properly transmitted to the Workspace object initialization.

NOTE: because the 'title' argument is now near-mandatory, it has
been bumped up to be the thrid positional argument to
create_general_report, shifting 'confidenceLevel' back to the 4th
argument.  This may cause BACKWARD COMPATIBILITY ISSUES, and so
a ValueError checks if the title is an int and raises an
explanatory ValueError to try to mitigate this.