Adds plotly hack to enable pickling plotly figures.
Plotly figures are unable to be pickled due to an inflexible
implementation of plotly.graph_objs.PlotlyDict.  This commit adds
`enable_plotly_pickling` and `disable_plotly_pickling` functions
which apply and repeal a hack to the internals of PlotlyDict that
enable pickling and un-pickling.  These functions are now sandwich
the pickling and un-pickling calls of a Workspace's cache, as
this is where plotly figure pickling normally occurs in pygsti.

Functionality has been tested using the Workspace Pickling example
notebook, and seems to be working as now the list of unpicklable
objects is empty (whereas before this commit it contained all
plotly-containing objects).  A sidenote: the test for setting the
'ws' member was changed from "hasattr(v, 'ws')" to
"isinstance(v, WorkspaceOutput)" since PlotlyDict objects say
they contain anything for some reason - so hasattr returns True
when we'd rather it didn't.