Removes the don't-add-zero-counts behavior from DataSet.
A DataSet's add_count_list method used to silently discard requests
to add data in which all the counts were zero.  This can be
problematic, especially in the case of intermediate measurements, where
it's OK for some gate sequences to have zero counts and they need to
be in the DataSet so they can be queried by other routines.  This commit
removes the problematic behavior and now a warning is generated if
zero-count data is added AND the DataSet has no measurement gate labels.