Adds new DatasetComparisonSummaryPlot to Workspace object.
This new plot type displays the #sigma deviance of the total-logL
scores between pairs of datasets given an appropriately formatted
dictionary of DataComparator objects.  This commit also renames
the old DatasetComparisionPlot class to DatasetComparisonHistogramPlot,
as this name is more descriptive.  The new plot has been added to
the general report.

As a side effect, the formatting function used to format values for
the boxes of matrix/box plots now explicitly converts NANs to
blanks.  I think this is always desirable, but may have unintented
side effects.