Adds eigenvalues of germs to table of gate eignvalues in reports.
This commit adds a new type of reportable quantity - one that takes
a gate set and gate string as arguments - and creates an eigenvalue
function that can be used to report the (gauge-invariant) eigenvalues
of the germs.  This output has been added to the table which
previously held just the eigenvalues of the individual gates.  The
table has been updated in report notebooks also.

Additionally, a column holding the eigenvalues of the *target*
 gates/germs has been added to this same table.