Adds error bars to a number of report tables/plots.
Attempts to enable error bar generation on everything possible
within the reports.  The one known exception is the spam table,
which because it does not show the Hilbert-Schmidt vector now,
does not show any error bars.

Note that error bars are shown in the hover labels of the
error generator & projections box-plot table.