Changes rpeconfig class's __init__ to be silent or raise a ValueError.
Previously rpeconfig.__init__ would print "Fully specified RPE configuation"
or a message indicating what parts of the configuration were missing.
This was slightly annoying in that several configs are initialized
upon importing the rpe sub-module and so statments would get printed
as a side effect of importing.  In addition to *not* printing
anything in the fully-specified case, the raising of a ValueError
 when the configuration is not fully specified seems more robust.